requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
httpx==0.23.0
nonebot-adapter-onebot==2.2.0
nonebot-plugin-apscheduler==0.2.0
nonebot2==2.0.0rc2
Pillow==9.2.0